cosmetic

+++ Src/CZICmd/BitmapGenGdiplus.cpp
@@ -21,60 +21,60 @@ using namespace libCZI;
 class CGdiplusBitmapWrapper : public libCZI::IBitmapData
 {
 private:
-	shared_ptr<Bitmap> bitmap;
-	BitmapData bd;
+    shared_ptr<Bitmap> bitmap;
+    BitmapData bd;
 public:
     CGdiplusBitmapWrapper(shared_ptr<Bitmap> bitmap) :bitmap(bitmap) {}
-	virtual ~CGdiplusBitmapWrapper() = default;
+    virtual ~CGdiplusBitmapWrapper() = default;
 
-	virtual libCZI::PixelType GetPixelType() const
-	{
-		switch (this->bitmap->GetPixelFormat())
-		{
-		case PixelFormat24bppRGB:
-			return libCZI::PixelType::Bgr24;
-		case PixelFormat16bppGrayScale:
-			return libCZI::PixelType::Gray16;
-		default:
-			return libCZI::PixelType::Invalid;
-		}
-	}
+    virtual libCZI::PixelType GetPixelType() const
+    {
+        switch (this->bitmap->GetPixelFormat())
+        {
+        case PixelFormat24bppRGB:
+            return libCZI::PixelType::Bgr24;
+        case PixelFormat16bppGrayScale:
+            return libCZI::PixelType::Gray16;
+        default:
+            return libCZI::PixelType::Invalid;
+        }
+    }
 
-	virtual libCZI::IntSize	GetSize() const
-	{
-		return libCZI::IntSize{ this->bitmap->GetWidth(), this->bitmap->GetHeight() };
-	}
+    virtual libCZI::IntSize	GetSize() const
+    {
+        return libCZI::IntSize{ this->bitmap->GetWidth(), this->bitmap->GetHeight() };
+    }
 
-	virtual libCZI::BitmapLockInfo	Lock()
-	{
-		auto gdiplusPxlFmt = this->bitmap->GetPixelFormat();
-		libCZI::BitmapLockInfo bitmapLockInfo;
-		Rect rect(0, 0, (INT)this->bitmap->GetWidth(), (INT)this->bitmap->GetHeight());
-		this->bitmap->LockBits(&rect, ImageLockModeRead, gdiplusPxlFmt, &this->bd);
-		bitmapLockInfo.ptrData = this->bd.Scan0;
-		bitmapLockInfo.ptrDataRoi = this->bd.Scan0;
-		bitmapLockInfo.stride = this->bd.Stride;
-        bitmapLockInfo.size = this->bd.Stride * this->bd.Height;
-		return bitmapLockInfo;
-	}
+    virtual libCZI::BitmapLockInfo Lock()
+    {
+        auto gdiplusPxlFmt = this->bitmap->GetPixelFormat();
+        libCZI::BitmapLockInfo bitmapLockInfo;
+        Rect rect(0, 0, (INT)this->bitmap->GetWidth(), (INT)this->bitmap->GetHeight());
+        this->bitmap->LockBits(&rect, ImageLockModeRead, gdiplusPxlFmt, &this->bd);
+        bitmapLockInfo.ptrData = this->bd.Scan0;
+        bitmapLockInfo.ptrDataRoi = this->bd.Scan0;
+        bitmapLockInfo.stride = this->bd.Stride;
+        bitmapLockInfo.size = static_cast<size_t>(this->bd.Stride) * this->bd.Height;
+        return bitmapLockInfo;
+    }
 
-	virtual void Unlock()
-	{
-		this->bitmap->UnlockBits(&this->bd);
-	}
+    virtual void Unlock()
+    {
+        this->bitmap->UnlockBits(&this->bd);
+    }
 };
 
 /*static*/ULONG_PTR CBitmapGenGdiplus::gdiplusToken;
 
 /*static*/void CBitmapGenGdiplus::Initialize()
 {
-	GdiplusStartupInput gdiplusStartupInput;
-	Status st = GdiplusStartup(&CBitmapGenGdiplus::gdiplusToken, &gdiplusStartupInput, NULL);
+    GdiplusStartupInput gdiplusStartupInput;
+    Status st = GdiplusStartup(&CBitmapGenGdiplus::gdiplusToken, &gdiplusStartupInput, NULL);
 }
 
 /*static*/void CBitmapGenGdiplus::Shutdown()
 {
-	GdiplusShutdown(CBitmapGenGdiplus::gdiplusToken);
+    GdiplusShutdown(CBitmapGenGdiplus::gdiplusToken);
 }
 
 CBitmapGenGdiplus::CBitmapGenGdiplus() : CBitmapGenGdiplus(nullptr)
@@ -89,15 +89,15 @@ CBitmapGenGdiplus::CBitmapGenGdiplus(const IBitmapGenParameters* params) : fonth
         this->fontheight = params->GetFontHeight();
     }
 
-	if (this->fontname.empty())
-	{
-		this->fontname = L"Arial Narrow";
-	}
+    if (this->fontname.empty())
+    {
+        this->fontname = L"Arial Narrow";
+    }
 
-	if (this->fontheight <= 0)
-	{
-		this->fontheight = 36;
-	}
+    if (this->fontheight <= 0)
+    {
+        this->fontheight = 36;
+    }
 }
 
 /*virtual*/CBitmapGenGdiplus::~CBitmapGenGdiplus()
@@ -106,69 +106,69 @@ CBitmapGenGdiplus::CBitmapGenGdiplus(const IBitmapGenParameters* params) : fonth
 
 /*virtual*/std::shared_ptr<libCZI::IBitmapData> CBitmapGenGdiplus::Create(libCZI::PixelType pixeltype, std::uint32_t width, std::uint32_t height, const BitmapGenInfo& info)
 {
-	PixelFormat pxlFmt;
-	switch (pixeltype)
-	{
-	case PixelType::Bgr24:
-	case PixelType::Gray16:
+    PixelFormat pxlFmt;
+    switch (pixeltype)
+    {
+    case PixelType::Bgr24:
+    case PixelType::Gray16:
     case PixelType::Bgr48:
     case PixelType::Gray8:
-		pxlFmt = PixelFormat24bppRGB;
-		break;
-	default:
-		throw std::exception("unsupported pixelformat");
-	}
+        pxlFmt = PixelFormat24bppRGB;
+        break;
+    default:
+        throw std::exception("unsupported pixelformat");
+    }
 
-	shared_ptr<Bitmap> bitmap = make_shared<Bitmap>(width, height, pxlFmt);
-	Graphics g(bitmap.get());
-	g.Clear(Color(255, 0, 0));
+    shared_ptr<Bitmap> bitmap = make_shared<Bitmap>(width, height, pxlFmt);
+    Graphics g(bitmap.get());
+    g.Clear(Color(255, 0, 0));
 
-	Font font(this->fontname.c_str(), (REAL)this->fontheight, FontStyle::FontStyleBold, UnitPoint);
-	SolidBrush brush(Color(0, 0, 0));
+    Font font(this->fontname.c_str(), (REAL)this->fontheight, FontStyle::FontStyleBold, UnitPoint);
+    SolidBrush brush(Color(0, 0, 0));
 
-	auto text = IBitmapGen::CreateTextW(info);
+    auto text = IBitmapGen::CreateTextW(info);
 
-	g.DrawString(
-		text.c_str(),
-		-1,
-		&font,
-		PointF(100, 300),
-		&brush);
+    g.DrawString(
+        text.c_str(),
+        -1,
+        &font,
+        PointF(100, 300),
+        &brush);
 
-	if (pixeltype == PixelType::Gray16)
-	{
-		auto bw = make_shared<CNullBitmapWrapper>(PixelType::Gray16, width, height);
-		Rect rect(0, 0, width, height);
-		BitmapData bd;
-		bitmap->LockBits(&rect, ImageLockModeRead, PixelFormat24bppRGB, &bd);
-		auto _ = finally([&] { bitmap->UnlockBits(&bd); });
+    if (pixeltype == PixelType::Gray16)
+    {
+        auto bw = make_shared<CNullBitmapWrapper>(PixelType::Gray16, width, height);
+        Rect rect(0, 0, width, height);
+        BitmapData bd;
+        bitmap->LockBits(&rect, ImageLockModeRead, PixelFormat24bppRGB, &bd);
+        auto _ = finally([&] { bitmap->UnlockBits(&bd); });
 
-		ScopedBitmapLockerSP dstBm{ bw };
-		for (uint32_t y = 0; y < height; ++y)
-		{
-			const std::uint8_t* src = static_cast<const std::uint8_t*>(bd.Scan0) + y * static_cast<size_t>(bd.Stride);
+        ScopedBitmapLockerSP dstBm{ bw };
+        for (uint32_t y = 0; y < height; ++y)
+        {
+            const std::uint8_t* src = static_cast<const std::uint8_t*>(bd.Scan0) + y * static_cast<size_t>(bd.Stride);
             std::uint16_t* dst = reinterpret_cast<std::uint16_t*>(static_cast<std::uint8_t*>(dstBm.ptrDataRoi) + y * static_cast<size_t>(dstBm.stride));
-			for (uint32_t x = 0; x < width; ++x)
-			{
-				uint8_t r = src[0];
-				uint8_t g = src[1];
-				uint8_t b = src[2];
-				uint16_t p;
-				if (r == 0 && g == 0 && b == 0)
-				{
-					p = 0;
-				}
-				else
-				{
-					p = 0xffff;
-				}
+            for (uint32_t x = 0; x < width; ++x)
+            {
+                uint8_t r = src[0];
+                uint8_t g = src[1];
+                uint8_t b = src[2];
+                uint16_t p;
+                if (r == 0 && g == 0 && b == 0)
+                {
+                    p = 0;
+                }
+                else
+                {
+                    p = 0xffff;
+                }
 
-				*dst = p;
+                *dst = p;
 
-				src += 3;
-				++dst;
-			}
-		}
+                src += 3;
+                ++dst;
+            }
+        }
 
         return bw;
     }
@@ -178,7 +178,7 @@ CBitmapGenGdiplus::CBitmapGenGdiplus(const IBitmapGenParameters* params) : fonth
         Rect rect(0, 0, width, height);
         BitmapData bd;
         bitmap->LockBits(&rect, ImageLockModeRead, PixelFormat24bppRGB, &bd);
-		auto _ = finally([&] { bitmap->UnlockBits(&bd); });	// execute "UnlockBits" when leaving scope
+        auto _ = finally([&] { bitmap->UnlockBits(&bd); });	// execute "UnlockBits" when leaving scope
 
         ScopedBitmapLockerSP dstBm{ bw };
         for (uint32_t y = 0; y < height; ++y)
@@ -215,7 +215,7 @@ CBitmapGenGdiplus::CBitmapGenGdiplus(const IBitmapGenParameters* params) : fonth
         Rect rect(0, 0, width, height);
         BitmapData bd;
         bitmap->LockBits(&rect, ImageLockModeRead, PixelFormat24bppRGB, &bd);
-		auto _ = finally([&] { bitmap->UnlockBits(&bd); });
+        auto _ = finally([&] { bitmap->UnlockBits(&bd); });
 
         ScopedBitmapLockerSP dstBm{ bw };
         for (uint32_t y = 0; y < height; ++y)
@@ -246,23 +246,23 @@ CBitmapGenGdiplus::CBitmapGenGdiplus(const IBitmapGenParameters* params) : fonth
             }
         }
 
-		return bw;
-	}
+        return bw;
+    }
 
-	return make_shared<CGdiplusBitmapWrapper>(bitmap);
+    return make_shared<CGdiplusBitmapWrapper>(bitmap);
 }
 
 /*static*/void CBitmapGenGdiplus::ConvertRgb24ToBgr24(std::uint32_t w, std::uint32_t h, std::uint32_t stride, void* ptrData)
 {
-	for (std::uint32_t y = 0; y < h; y++)
-	{
-		std::uint8_t* p = static_cast<std::uint8_t*>(ptrData) + y * static_cast<size_t>(stride);
-		for (std::uint32_t x = 0; x < w; x++)
-		{
-			std::swap(*p, *(p + 2));
-			p += 3;
-		}
-	}
+    for (std::uint32_t y = 0; y < h; y++)
+    {
+        std::uint8_t* p = static_cast<std::uint8_t*>(ptrData) + y * static_cast<size_t>(stride);
+        for (std::uint32_t x = 0; x < w; x++)
+        {
+            std::swap(*p, *(p + 2));
+            p += 3;
+        }
+    }
 }
 
 #endif
